Isolate the different styles
