+++ build.gradle
@@ -15,7 +15,7 @@ buildscript {
         classpath "org.junit.platform:junit-platform-gradle-plugin:$junitPlatformVersion"
         classpath 'org.asciidoctor:asciidoctor-gradle-plugin:1.5.3'
         classpath 'org.ajoberstar:gradle-git:1.7.2'
-        classpath "gradle.plugin.org.jetbrains.intellij.plugins:gradle-intellij-plugin:0.3.1"
+        classpath "gradle.plugin.org.jetbrains.intellij.plugins:gradle-intellij-plugin:0.3.4"
         classpath "org.jfrog.buildinfo:build-info-extractor-gradle:4.6.0"
         classpath "com.github.jengelman.gradle.plugins:shadow:2.0.3"
     }
