+++ gradle.properties
@@ -1,4 +1,4 @@
-kotlinVersion=1.2.41
+kotlinVersion=1.2.50
 junitPlatformVersion=1.1.1
 kotlin.incremental=true
 githubRepo=https://github.com/spekframework/spek
