+++ gradle/common/dependencies.gradle
@@ -27,6 +27,11 @@ dependencyManagement {
 
         dependency 'com.xenomachina:kotlin-argparser:2.0.6'
 
-        dependency "com.nhaarman:mockito-kotlin:1.5.0"
+        dependency "com.nhaarman.mockitokotlin2:mockito-kotlin:2.0.0-RC1"
+
+        dependencySet(group: 'org.mockito', version: '2.19.0') {
+            entry 'mockito-core'
+            entry 'mockito-inline'
+        }
     }
 }
