+++ integration-test/common/src/test/kotlin/org/spekframework/spek2/integration/SetFeature.kt
@@ -2,6 +2,7 @@ package org.spekframework.spek2.integration
 
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.gherkin.Feature
+import org.spekframework.spek2.style.specification.describe
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
