--- spek-dsl/common/src/main/kotlin/org/spekframework/spek2/dsl/ScopeBody.kt
@@ -1,7 +1,0 @@
-package org.spekframework.spek2.dsl
-
-import org.spekframework.spek2.lifecycle.MemoizedValue
-
-interface ScopeBody {
-    fun <T> memoized(): MemoizedValue<T>
-}
