--- spek-dsl/common/src/main/kotlin/org/spekframework/spek2/dsl/Skip.kt
@@ -1,6 +1,0 @@
-package org.spekframework.spek2.dsl
-
-sealed class Skip {
-    class Yes(val reason: String? = null) : Skip()
-    object No : Skip()
-}
