--- spek-dsl/common/src/main/kotlin/org/spekframework/spek2/dsl/TestBody.kt
@@ -1,6 +1,0 @@
-package org.spekframework.spek2.dsl
-
-import org.spekframework.spek2.meta.SpekDsl
-
-@SpekDsl
-interface TestBody : ScopeBody
