--- spek-dsl/common/src/main/kotlin/org/spekframework/spek2/dsl/TestContainer.kt
@@ -1,10 +1,0 @@
-package org.spekframework.spek2.dsl
-
-import org.spekframework.spek2.meta.*
-
-@SpekDsl
-interface TestContainer {
-    @Synonym(type = SynonymType.TEST)
-    @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
-    fun test(description: String, skip: Skip = Skip.No, body: TestBody.() -> Unit)
-}
