+++ spek-dsl/common/src/main/kotlin/org/spekframework/spek2/style/specification/specificationStyle.kt
@@ -1,48 +1,49 @@
 package org.spekframework.spek2.style.specification
 
 import org.spekframework.spek2.dsl.GroupBody
+import org.spekframework.spek2.dsl.LifecycleAware
 import org.spekframework.spek2.dsl.Skip
 import org.spekframework.spek2.dsl.TestBody
 import org.spekframework.spek2.lifecycle.CachingMode
 import org.spekframework.spek2.meta.*
 
 @SpekDsl
-class Suite(delegate: GroupBody) : GroupBody by delegate {
+class Suite(val delegate: GroupBody) : LifecycleAware by delegate {
 
     @Synonym(SynonymType.GROUP)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun describe(description: String, skip: Skip = Skip.No, body: Suite.() -> Unit) {
-        createSuite(description, skip, body)
+        delegate.createSuite(description, skip, body)
     }
 
     @Synonym(SynonymType.GROUP)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun context(description: String, skip: Skip = Skip.No, body: Suite.() -> Unit) {
-        createSuite(description, skip, body)
+        delegate.createSuite(description, skip, body)
     }
 
     @Synonym(SynonymType.GROUP, excluded = true)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun xdescribe(description: String, reason: String = "", body: Suite.() -> Unit) {
-        createSuite(description, Skip.Yes(reason), body)
+        delegate.createSuite(description, Skip.Yes(reason), body)
     }
 
     @Synonym(SynonymType.GROUP, excluded = true)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun xcontext(description: String, reason: String = "", body: Suite.() -> Unit) {
-        createSuite(description, Skip.Yes(reason), body)
+        delegate.createSuite(description, Skip.Yes(reason), body)
     }
 
     @Synonym(SynonymType.TEST)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun it(description: String, skip: Skip = Skip.No, body: TestBody.() -> Unit) {
-        createTest(description, skip, body)
+        delegate.createTest(description, skip, body)
     }
 
     @Synonym(SynonymType.TEST, excluded = true)
     @Descriptions(Description(DescriptionLocation.VALUE_PARAMETER, 0))
     fun xit(description: String, reason: String = "", body: TestBody.() -> Unit) {
-        createTest(description, Skip.Yes(reason), body)
+        delegate.createTest(description, Skip.Yes(reason), body)
     }
 
     fun before(cb: () -> Unit) {
