+++ spek-runner/junit5/src/test/kotlin/org/spekframework/spek2/junit/SpekTestDescriptorFactoryTest.kt
@@ -2,7 +2,7 @@ package org.spekframework.spek2.junit
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.sameInstance
-import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockitokotlin2.mock
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.spekframework.spek2.dsl.Skip
